freshen code as singleton class
